requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==45.0.5
Deprecated==1.2.18
Django==5.2.4
django-allauth==65.10.0
django-audit-log==0.7.0
django-axes==8.0.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
django-formtools==2.5.1
django-guardian==3.0.3
django-ipware==7.0.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-prometheus==2.4.1
django-ratelimit==4.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
line-bot-sdk==3.17.1
multidict==6.6.3
packaging==25.0
phonenumbers==9.0.10
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
qrcode==7.4.2
redis==6.2.0
requests==2.32.4
sentry-sdk==2.34.1
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.1
